test: extract RuleCommands; run good/rules against real iptables
Move the good/rules command sequence out of GoodTests into a
driver-agnostic RuleCommands (tests/functional/commands.py) so it
can run two ways:

- GoodTests(RuleCommands, FunctionalTestCase): in-process, fake
  iptables, transcript + render oracle (unchanged behavior)
- GoodE2E(RuleCommands, E2ETestCase): subprocess, REAL iptables, ufw
  enabled and --dry-run stripped, so every assert_ok is an
  apply-acceptance check against real iptables/ip6tables-restore

The on-disk generation checks (assert_rule_count etc.) are gated by
verify_on_disk: real in the functest, no-ops in e2e.
